pcp-atop: extend QA with archives containing nfs metrics
Improve pcp-atop handling of metrics missing from archives, and
correct a cut+paste error in the recently added NFS metric value
extraction code.  Document that pcp-atop will optionally use the
metrics of pmdaapache(1) and pmdanfsclient(1).

+++ qa/archives/mk.atop.sh
@@ -28,6 +28,12 @@ then
     exit 1
 fi
 
+if pmprobe nfsclient >&1 | grep -q 'Unknown metric name'
+then
+    echo "Arrgh, nfsclient PMDA is apparently not installed"
+    exit 1
+fi
+
 trap "rm -rf $tmp; exit" 0 1 2 3 15
 
 mkdir -p $tmp/config
